[Go] move logging out of internal
Move logging support out of internal so third-party plugins
can use it.

We can't put it in core because tracing needs it, so I made a
package for it. At that point I felt that `logger.FromContext`
read better than `logging.Logger`.
