+++ go/core/action.go
@@ -22,8 +22,9 @@ import (
 	"reflect"
 	"time"
 
-	"github.com/firebase/genkit/go/internal"
+	"github.com/firebase/genkit/go/core/logger"
 	"github.com/firebase/genkit/go/core/tracing"
+	"github.com/firebase/genkit/go/internal"
 	"github.com/invopop/jsonschema"
 )
 
@@ -102,11 +103,11 @@ func (a *Action[I, O, S]) setTracingState(tstate *tracing.State) { a.tstate = ts
 func (a *Action[I, O, S]) Run(ctx context.Context, input I, cb func(context.Context, S) error) (output O, err error) {
 	// TODO: validate input against JSONSchema for I.
 	// TODO: validate output against JSONSchema for O.
-	internal.Logger(ctx).Debug("Action.Run",
+	logger.FromContext(ctx).Debug("Action.Run",
 		"name", a.name,
 		"input", fmt.Sprintf("%#v", input))
 	defer func() {
-		internal.Logger(ctx).Debug("Action.Run",
+		logger.FromContext(ctx).Debug("Action.Run",
 			"name", a.name,
 			"output", fmt.Sprintf("%#v", output),
 			"err", err)
